Add simple harpsichord synth
I've added a simple harpsichord synth, based on Karplus-Strong string
synthesis. The main magic happens in the #'pluck filter - it is a
resonant filter, with resonant frequency controlled by the freq
parameter by changing the delay length.

You can control the brightness of the harpsichord by changing the
one-pole coefficient (the last parameter of the pluck filter). Values
closer to 0 are brighter, but harsher; while values closer to 1 are more
mellow, but duller. In fact, it's fun to play with this parameter while
the tune is playing :)